requirements.txt
openai==1.35.14
python-dotenv==1.0.1